Removed redundant base service class.
